вносит изменения в структуру проекта
