+++ src/components/locations-list/locations-list.tsx
@@ -1,6 +1,5 @@
+import LocationTab from '../../components/location-tab/location-tab';
 import {CITIES} from '../../consts';
-import {changeCity} from '../../store/app-process-slice/app-process-slice';
-import {useAppDispatch} from '../../hooks/index';
 import {OfferCity} from '../../types/offers-types';
 
 type currentLocationProps = {
@@ -8,24 +7,16 @@ type currentLocationProps = {
 }
 
 function LocationsList({currentLocation}: currentLocationProps): JSX.Element {
-  const dispatch = useAppDispatch();
   return (
     <div className="tabs">
       <section className="locations container">
         <ul className="locations__list tabs__list">
           {CITIES.map((city) => (
-            <li key={city.name} className="locations__item">
-              <a
-                className={`locations__item-link tabs__item ${city.name === currentLocation.name ? 'tabs__item--active' : ''}`}
-                href="#"
-                onClick={(evt) => {
-                  evt.preventDefault();
-                  dispatch(changeCity(city));
-                }}
-              >
-                <span>{city.name}</span>
-              </a>
-            </li>
+            <LocationTab
+              key={city.name}
+              currentLocation={currentLocation}
+              city={city}
+            />
           ))}
         </ul>
       </section>
